preload/preload-strict-dynamic.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js" nonce="123"></script>
<script src="/resources/testharnessreport.js" nonce="123"></script>
<title>CSP strict-dynamic + preload</title>
<meta http-equiv="Content-Security-Policy" content="script-src 'nonce-123' 'strict-dynamic'" />
</head>
<body>
<link id="static-no-nonce" href="resources/dummy.js?static-no-nonce" rel=preload as=script>
<link id="static-nonce" href="resources/dummy.js?static-nonce" rel=preload as=script nonce="123">
<script nonce="123">
  let isLoaded = (url) => {
    let entries = performance.getEntriesByType("resource");
    for (let entry of entries) {
      if (entry.name.indexOf(url) != -1 ) {
        return true;
      }
    }
    return false;
  }
  let link = document.createElement("link");
  link.rel = "preload";
  link.href = "resources/dummy.js?dynamic-nonce";
  link.as = "script";
  document.head.appendChild(link);
  var t = async_test('preload from nonced script should work with strict-dynamic. preloaded script from markup should not.');
  t.step_timeout(() => {
    assert_true(isLoaded("dynamic-nonce"), "dynamic nonce load");
    assert_true(isLoaded("static-nonce"), "static nonce load");
    assert_false(isLoaded("static-no-nonce"), "static no nonce error");
    t.done();
  }, 500);

</script>
</body>
</html>
